refactor: move agents model from service to model
