+++ hive-api/routes.rest
@@ -15,8 +15,8 @@ POST http://{{host}}/api/agents HTTP/1.1
 content-type: application/json
 
 {
-  "name": "PikaChu Y", 
-  "email": "pikachu.y@gmail.com"
+  "name": "PikaChu YY", 
+  "email": "pikachu.yy@gmail.com"
 }
 
 ### UPDATE agent
@@ -24,9 +24,9 @@ PUT   http://{{host}}/api/agents/ HTTP/1.1
 content-type: application/json
 
 {
-  "id": 7,
-  "name": "PikaChuYY", 
-  "email": "pikachu.chiu.yy@gmail.com"
+  "id": 1,
+  "name": "PikaChuY", 
+  "email": "pikachu.chiu.y@gmail.com"
 }
 
 ### DELETE agent
@@ -34,7 +34,7 @@ DELETE http://{{host}}/api/agents/ HTTP/1.1
 content-type: application/json
 
 {
-  "id": 7
+  "id": 1
 }
 
 ###### POKEMON API ######
