+++ hive-api/src/interfaces/agent.interface.ts
@@ -1,0 +1,4 @@
+export interface IAgent {
+  name: string;
+  email: string;
+}
